# openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
...
    "/change_password": {
      "post": {
        "tags": ["auth"],
        "summary": "Change Password",
        "operationId": "change_password_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/fetch_items": {
      "get": {
        "tags": ["core"],
        "summary": "Fetch Items",
        "operationId": "fetch_items_fetch_items_get",
        "parameters": [
          {
            "name": "only_wish_list",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Only Wish List"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Sortorder"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "title": "Limit"
            }
          },
          {
            "name": "searchQuery",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Searchquery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FetchItemsResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/add_ean_to_list/": {
      "post": {
        "tags": ["core"],
        "summary": "Add Ean To List",
        "operationId": "add_ean_to_list_add_ean_to_list__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddEanRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/add_fetched_items": {
      "post": {
        "tags": ["core"],
        "summary": "Add Fetched Items",
        "operationId": "add_fetched_items_add_fetched_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddFetchedItems" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/get_catalogue_offers": {
      "get": {
        "tags": ["geo"],
        "summary": "Get Offers",
        "operationId": "get_offers_get_catalogue_offers_get",
        "parameters": [
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "schema": { "type": "number", "title": "Longitude" }
          },
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "schema": { "type": "number", "title": "Latitude" }
          },
          {
            "name": "DeprecationDays",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "title": "Deprecationdays"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/add_new_market": {
      "post": {
        "tags": ["geo"],
        "summary": "Post Market",
        "operationId": "post_market_add_new_market_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddSupermarketRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/post_catalogue": {
      "post": {
        "tags": ["geo"],
        "summary": "Post Catalogue",
        "operationId": "post_catalogue_post_catalogue_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_catalogue_post_catalogue_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/get_supermarkets_close": {
      "get": {
        "tags": ["geo"],
        "summary": "Get Supermarkets Close",
        "operationId": "get_supermarkets_close_get_supermarkets_close_get",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "number" }, { "type": "null" }],
              "title": "Lat"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "number" }, { "type": "null" }],
              "title": "Lon"
            }
          },
          {
            "name": "radius_meters",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50000,
              "minimum": 100,
              "default": 2000,
              "title": "Radius Meters"
            }
          },
          {
            "name": "chains",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "array", "items": { "type": "string" } },
                { "type": "null" }
              ],
              "title": "Chains"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/recipes": {
      "get": {
        "tags": ["meal planning"],
        "summary": "Get Recipes",
        "operationId": "get_recipes_recipes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        }
      },
      "post": {
        "tags": ["meal planning"],
        "summary": "Add Recipe",
        "operationId": "add_recipe_recipes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddRecipe" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/planner_settings": {
      "get": {
        "tags": ["meal planning"],
        "summary": "Get Planner Settings",
        "operationId": "get_planner_settings_planner_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        }
      },
      "post": {
        "tags": ["meal planning"],
        "summary": "Put Planner Settings",
        "operationId": "put_planner_settings_planner_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PlannerSettings" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/week_plan": {
      "get": {
        "tags": ["meal planning"],
        "summary": "Get Week Plan",
        "operationId": "get_week_plan_week_plan_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        }
      },
      "post": {
        "tags": ["meal planning"],
        "summary": "Post Week Plan",
        "operationId": "post_week_plan_week_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddWeekPlan" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["meal planning"],
        "summary": "Delete Meal Slot",
        "operationId": "delete_meal_slot_week_plan_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MealSlot" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/week_plan/replace_meal": {
      "post": {
        "tags": ["meal planning"],
        "summary": "Replace Meal",
        "operationId": "replace_meal_week_plan_replace_meal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MealSlot" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddEanRequest": {
        "properties": {
          "ean": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Ean"
          },
          "item_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Item Name"
          },
          "count": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Count",
            "default": 1
          },
          "wish_list": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Wish List"
          },
          "quantity_data": {
            "anyOf": [
              { "$ref": "#/components/schemas/QuantityInfo" },
              { "type": "null" }
            ]
          }
        },
        "type": "object",
        "title": "AddEanRequest",
        "description": "Request model for add_ean_to_list endpoint"
      },
      "AddFetchedItems": {
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/FetchedSingleItem" },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": ["items"],
        "title": "AddFetchedItems"
      },
      "AddRecipe": {
        "properties": {
          "name": { "type": "string", "title": "Name" },
          "emoji": { "type": "string", "title": "Emoji" },
          "baseTime": { "type": "integer", "title": "Basetime" },
          "baseServings": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Baseservings",
            "default": 2
          },
          "tags": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Tags",
            "default": []
          },
          "favorited": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Favorited",
            "default": false
          },
          "ingredients": {
            "items": { "$ref": "#/components/schemas/Ingredient" },
            "type": "array",
            "title": "Ingredients"
          },
          "steps": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Steps"
          }
        },
        "type": "object",
        "required": ["name", "emoji", "baseTime", "ingredients", "steps"],
        "title": "AddRecipe"
      },
      "AddSupermarketRequest": {
        "properties": {
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "address": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address"
          },
          "city": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "City"
          },
          "category": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Category"
          },
          "postcode": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Postcode"
          },
          "latitude": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Longitude"
          }
        },
        "type": "object",
        "title": "AddSupermarketRequest"
      },
      "AddWeekPlan": {
        "properties": {
          "week": { "$ref": "#/components/schemas/WeekPlan" },
          "DaySettings": { "$ref": "#/components/schemas/WeekSettings" }
        },
        "type": "object",
        "required": ["week", "DaySettings"],
        "title": "AddWeekPlan"
      },
...
      "Body_post_catalogue_post_catalogue_post": {
        "properties": {
          "catalogue": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Catalogue"
          },
          "latitude": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Longitude"
          },
          "lat": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Lon"
          },
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "address": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Address"
          },
          "categories": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Categories"
          },
          "city": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "City"
          },
          "postcode": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Postcode"
          }
        },
        "type": "object",
        "required": ["catalogue"],
        "title": "Body_post_catalogue_post_catalogue_post"
      },
      "ChangePasswordRequest": {
        "properties": {
          "current_password": { "type": "string", "title": "Current Password" },
          "new_password": { "type": "string", "title": "New Password" }
        },
        "type": "object",
        "required": ["current_password", "new_password"],
        "title": "ChangePasswordRequest"
      },
...
      "FetchItemsResponse": {
        "properties": {
          "status": { "type": "integer", "const": 200, "title": "Status" },
          "items": {
            "items": { "$ref": "#/components/schemas/ItemsFetched" },
            "type": "array",
            "title": "Items"
          },
          "distinct_items": { "type": "integer", "title": "Distinct Items" },
          "accumulated_count": {
            "type": "integer",
            "title": "Accumulated Count"
          }
        },
        "type": "object",
        "required": ["status", "items", "distinct_items", "accumulated_count"],
        "title": "FetchItemsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MealSlot": {
        "properties": {
          "recipe_id": { "type": "integer", "title": "Recipe Id" },
          "servings": { "type": "integer", "title": "Servings" },
          "meal_type": { "type": "string", "title": "Meal Type" },
          "day": { "type": "string", "title": "Day" },
          "day_time": { "type": "string", "title": "Day Time" }
        },
        "type": "object",
        "required": ["recipe_id", "servings", "meal_type", "day", "day_time"],
        "title": "MealSlot"
      },
      "PlannerSettings": {
        "properties": {
          "defaultServings": { "type": "integer", "title": "Defaultservings" },
          "quickMealMinutes": {
            "type": "integer",
            "title": "Quickmealminutes"
          },
          "normalMealMinutes": {
            "type": "integer",
            "title": "Normalmealminutes"
          }
        },
        "type": "object",
        "required": [
          "defaultServings",
          "quickMealMinutes",
          "normalMealMinutes"
        ],
        "title": "PlannerSettings"
      },
      "QuantityInfo": {
        "properties": {
          "product_quantity": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Product Quantity"
          },
          "product_quantity_unit": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Product Quantity Unit"
          }
        },
        "type": "object",
        "required": ["product_quantity", "product_quantity_unit"],
        "title": "QuantityInfo"
      },
...
    }
  }
}